requirements.txt
numpy
matplotlib

torch
torchvision
pillow==6.2.1



#Tutorial Requirements
#matplotlib==2.0.0
#pandas==0.19.2
#scikit-learn==0.18.1
#scipy==1.0.0